permissions

+++ command.py
@@ -56,16 +56,24 @@ class KillCommand(Command):
 
     async def execute(self, message):
         username = self.get_args(message)
+        server_id = message.guild.id
+        if not db.can_kill(str(server_id)):
+            await channel.send('No.')
+            return
         delete_jobs = []
-        for channel in message.guild.channels:
-            async for prev_message in channel.history():
+        for text_channel in message.guild.channels:
+            if not isinstance(text_channel, discord.TextChannel):
+                pass
+            async for prev_message in text_channel.history():
                 name = prev_message.author.name
                 discriminator = prev_message.author.discriminator
                 #TODO Verify This Works
                 if f'{name}#{discriminator}' == username:
                     delete_jobs.append(prev_message.delete())
         channel = message.channel
-        await asyncio.gather(message.delete(), *delete_jobs)
+        await asyncio.gather(message.delete(), 
+                             db.disable_kill(server_id),
+                             *delete_jobs)
         await channel.send(file=discord.File('assets/hando.jpg'))
         await channel.send(file=discord.File('assets/ideletedthisuser.jpg'))
